refactor(clawdstrike-policy-event): extract kinds/observation from edr/receipt/mod.rs
