+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -1,6 +1,7 @@
 mod detection;
 mod evidence_bundle;
 mod graph_slice;
+mod observation;
 mod policy_decision;
 mod policy_delta;
 mod policy_event;
@@ -12,6 +13,7 @@ mod simulation;
 pub(crate) use detection::*;
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
+pub(crate) use observation::*;
 pub(crate) use policy_decision::*;
 pub(crate) use policy_delta::*;
 pub(crate) use policy_event::*;
